requirements.txt
asgiref==3.8.1
Django==5.0.7
django-admin-notification==0.2
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-solo==2.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0